1) Cycle Detection Algorithm for Directed Graphs implemented. 2) Cycle Detection Algorithm for Un - Directed Graph improved.
Basically respective algorithms are for respective type of Graphs, have to see if there can be a generic algorithm for both Directed and Un-Directed.

But Generic Algorithm cannot be implemented, as let's consider there are 2 nodes [0, 1] and they are inter-connected, now in the case of Un-Directed Algorithm, the Graph doesn't have a Cycle, but in the case of Directed Algorithm, the Graph has a Cycle.

It is a Confusion !